[chromedriver] Fix issue with killing chrome after last window is closed.
NOTRY=true


Review URL: https://chromiumcodereview.appspot.com/12318026

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@183830 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 66a4213717da7989915ba0950f594bb49c8c7a25

+++ chrome/test/chromedriver/chrome_desktop_impl.cc
@@ -64,8 +64,12 @@ Status ChromeDesktopImpl::Launch(const base::FilePath& chrome_exe,
 }
 
 Status ChromeDesktopImpl::Quit() {
-  if (!base::KillProcess(process_, 0, true))
-    return Status(kUnknownError, "cannot kill Chrome");
+  if (!base::KillProcess(process_, 0, true)) {
+    int exit_code;
+    if (base::GetTerminationStatus(process_, &exit_code) ==
+        base::TERMINATION_STATUS_STILL_RUNNING)
+      return Status(kUnknownError, "cannot kill Chrome");
+  }
   return Status(kOk);
 }
